move apilib
