--- api/apilib/accesslog.go
@@ -1,24 +1,0 @@
-/*
-   Copyright (c) 2016 VMware, Inc. All Rights Reserved.
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-
-package apilib
-
-//AccessLog for go test
-type AccessLog struct {
-	Username       string `json:"username,omitempty"`
-	Keywords       string `json:"keywords,omitempty"`
-	BeginTimestamp int32  `json:"beginTimestamp,omitempty"`
-	EndTimestamp   int32  `json:"endTimestamp,omitempty"`
-}
