--- api/apilib/role.go
@@ -1,8 +1,0 @@
-package apilib
-
-//Role ...
-type Role struct {
-	RoleID   int32  `json:"role_id,omitempty"`
-	RoleCode string `json:"role_code,omitempty"`
-	RoleName string `json:"role_name,omitempty"`
-}
