--- api/apilib/user.go
@@ -1,12 +1,0 @@
-package apilib
-
-//User ...
-type User struct {
-	UserID   int32  `json:"user_id,omitempty"`
-	Username string `json:"username,omitempty"`
-	Email    string `json:"email,omitempty"`
-	Password string `json:"password,omitempty"`
-	Realname string `json:"realname,omitempty"`
-	Comment  string `json:"comment,omitempty"`
-	Deleted  int32  `json:"deleted,omitempty"`
-}
