+++ tests/apitests/apilib/projecttemp4search.go
@@ -1,9 +1,8 @@
-package HarborAPI
+package apilib
 
-import ()
-
+//Project4Search ...
 type Project4Search struct {
-	ProjectId    int32  `json:"id,omitempty"`
+	ProjectID    int32  `json:"id,omitempty"`
 	ProjectName  string `json:"name,omitempty"`
 	Public       int32   `json:"public,omitempty"`
 }
