+++ tests/apitests/apilib/repositorytemp4search.go
@@ -1,7 +1,10 @@
-package HarborAPI
+package apilib
 
+import ()
+
+//Repository4Search ...
 type Repository4Search struct {
-	ProjectId     int32     `json:"project_id,omitempty"`
+	ProjectID     int32     `json:"project_id,omitempty"`
 	ProjectName   string    `json:"project_name,omitempty"`
 	ProjectPublic int32     `json:"project_public,omitempty"`
 	RepoName      string    `json:"repository_name,omitempty"`
