+++ tests/apitests/apilib/role.go
@@ -1,7 +1,8 @@
-package HarborAPI
+package apilib
 
+//Role ...
 type Role struct {
-	RoleId   int32  `json:"role_id,omitempty"`
+	RoleID   int32  `json:"role_id,omitempty"`
 	RoleCode string `json:"role_code,omitempty"`
 	RoleName string `json:"role_name,omitempty"`
 }
