+++ tests/apitests/apilib/roleparam.go
@@ -1,5 +1,6 @@
-package HarborAPI
+package apilib
 
+//RoleParam ...
 type RoleParam struct {
 	Roles    []int32 `json:"roles,omitempty"`
 	UserName string  `json:"user_name,omitempty"`
