+++ tests/apitests/apilib/search.go
@@ -1,7 +1,8 @@
-package HarborAPI
+package apilib
 
 import ()
 
+//Search ...
 type Search struct {
 	Projects     []Project4Search    `json:"project,omitempty"`
 	Repositories []Repository4Search `json:"repository,omitempty"`
